+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Migrate formatter to newly created language server, which should help reduce delays.
 - Fix commad args for Bruno CLI when running requests with multiple included/excluded tags.
 - Make functions for determining test results per descendant item more resilient by avoiding fetching data from filesystem.
+- Increase resilience for function that should reveal item in collection explorer by improving error handling.
 
 ## [0.20.3] (2026-01-27)